+++ src/sections/IdentifierDoi.vue
@@ -49,8 +49,15 @@ watch(selected, (value) => {
 </script>
 
 <template>
-  <FormLabel>
-    <input type="checkbox" name="identifiers" :id="'identifier-doi-selection'" v-model="selected" />
-    DOI (NERC prefix)
-  </FormLabel>
+  <div>
+    <FormLabel>
+      <input
+        type="checkbox"
+        name="identifiers"
+        :id="'identifier-doi-selection'"
+        v-model="selected"
+      />
+      DOI (NERC prefix)
+    </FormLabel>
+  </div>
 </template>