+++ src/sections/IdentifierSelf.vue
@@ -39,15 +39,17 @@ watch(
 </script>
 
 <template>
-  <FormLabel>
-    <input
-      class="cursor-not-allowed"
-      type="checkbox"
-      name="identifiers"
-      :id="'identifier-self-selection'"
-      disabled
-      v-model="selected"
-    />
-    BAS Data Catalogue (<em>required</em>)
-  </FormLabel>
+  <div>
+    <FormLabel>
+      <input
+        class="cursor-not-allowed"
+        type="checkbox"
+        name="identifiers"
+        :id="'identifier-self-selection'"
+        disabled
+        v-model="selected"
+      />
+      BAS Data Catalogue (<em>required</em>)
+    </FormLabel>
+  </div>
 </template>